breaking: move bit-dependencies command to Harmony, add dependencies --debug as a new subcommand
